requirements.txt
asn1crypto==0.24.0
cachetools==4.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dialogflow==0.7.2
entrypoints==0.3
flake8==3.7.9
future==0.17.1
google-api-core==1.16.0
google-auth==1.11.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
mccabe==0.6.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-telegram-bot==11.1.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.8
